+++ src/managers/account_manager.h
@@ -48,7 +48,8 @@ typedef struct {
 
 typedef struct {
     uint8_t accountIndex;
-    char ufvkCache[ZCASH_UFVK_MAX_LEN + 1];
+    char mainnetUfvkCache[ZCASH_UFVK_MAX_LEN + 1];
+    char testnetUfvkCache[ZCASH_UFVK_MAX_LEN + 1];
     uint8_t seedFingerprint[32];
 } ZcashUFVKCache_t;
 
@@ -107,6 +108,7 @@ void AccountsDataCheck(void);
 bool IsZcashSupportedForCurrentMnemonic(void);
 bool IsMoneroSupportedForCurrentMnemonic(void);
 int32_t GetZcashUFVK(uint8_t accountIndex, char* outUFVK);
+int32_t GetZcashTestnetUFVK(uint8_t accountIndex, char* outUFVK);
 int32_t GetZcashSFP(uint8_t accountIndex, uint8_t* outSFP);
 int32_t SetupZcashSFP(uint8_t accountIndex, const char* password);
 #ifdef CYPHERPUNK_VERSION